raspberry/requirements.txt
spidev
mfrc522
raspberrypi-tm1637
RPi.GPIO
requests
